requirements.txt
streamlit>=1.38.0
PyPDF2>=3.0.1
python-dotenv>=1.0.1
langchain-groq>=0.1.3

